Move SoundOption component into a separate file
