+++ src/components/SegmentForm/index.tsx
@@ -2,11 +2,10 @@
 
 import {
   Music4Icon,
-  PauseIcon,
-  PlayIcon,
   PlusIcon,
   Trash2Icon,
 } from "lucide-react";
+
 import { useEffect, useRef, useState } from "react";
 
 import Select from "@/lib/ui/components/Select";
@@ -17,6 +16,7 @@ import Button from "@/lib/ui/components/Button";
 import SoundSettings from "../MuteSoundSwitch";
 import AddTagsInput from "../AddTagsInput";
 import SoundsPicker from "../SoundsPicker";
+import SoundOption from "../SoundOption";
 
 
 const SegmentForm = ({ segmentId }: { segmentId: Id }) => {
@@ -122,58 +122,6 @@ const SegmentSounds = ({ segment }: { segment: Segment }) => {
       />
 
       <AddTagsInput entity={segment} />
-    </div>
-  );
-};
-
-const SoundOption = ({ src }: { src: Sound["src"] }) => {
-  const [isPlaying, setIsPlaying] = useState(false);
-  const audio = useRef(new Audio(src));
-
-  useEffect(() => {
-    const onEnded = () => {
-      setIsPlaying(false);
-    };
-    audio.current.addEventListener("ended", onEnded);
-    return () => {
-      audio.current.removeEventListener("ended", onEnded);
-    };
-  }, []);
-
-  useEffect(() => {
-    if (isPlaying) audio.current.play();
-    else audio.current.pause();
-  }, [isPlaying]);
-
-  return (
-    <div className="flex overflow-hidden justify-between items-center">
-      {/* <span>{src.split("/").at(-1)}</span> */}
-      <span
-        className="whitespace-nowrap overflow-hidden text-ellipsis flex-1 text-start "
-        title={src}
-      >
-        {src.replace("/sounds", "")}
-      </span>
-
-      {isPlaying ? (
-        <PauseIcon
-          size={16}
-          strokeWidth={1}
-          onClick={(e) => {
-            e.stopPropagation();
-            setIsPlaying(false);
-          }}
-        />
-      ) : (
-        <PlayIcon
-          size={16}
-          strokeWidth={1}
-          onClick={(e) => {
-            e.stopPropagation();
-            setIsPlaying(true);
-          }}
-        />
-      )}
     </div>
   );
 };
